--- src/screens/invoice-builder/components/section/general-info.tsx
@@ -1,9 +1,0 @@
-import SectionContainer from "./section-container";
-
-export default function GeneralInfo() {
-  return (
-    <SectionContainer>
-      <div>test</div>
-    </SectionContainer>
-  );
-}
